requirements.txt
numpy>=1.26
pandas>=2.2
matplotlib>=3.8
scikit-learn>=1.4
python-dateutil>=2.9
portion>=2.4
optuna>=3.6
torch>=2.2

# ESA-ADB metrics: use the vendored ./timeeval package (do not pip install timeeval into the same env).
